Adding instructions for reviewers on PR template
We need to make it easy for non-developer reviewers (for example POs or URs) to
review the changes in PRs.

This last section can include a link to the exact place where the change can be
viewed and also instructions on how to get to that page with the right test
email or any other change/selection you have to do to view it.