Add a new endpoint for receiving "participant started" events
* Adds a new endpoint at `/api/v1/early_career_teacher_participants` that allows POSTing a single "participant started" event.
* The state of a participant is managed with the AASM gem (state machine) and stored in the `participation_record` table.
* The history of the `participation_record` is stored in the `participant_event` table by use of the `paper_trail` gem.
* The new endpoint uses bearer-token based authentication of lead providers with tokens stored in new table `lead_provider_api_tokens`.
* Adds a rake task to generate API token for lead provider
* Aliased `current_lead_provider` to `current_user` for papertrail whodunnit
* Add aasm-diagram generation rake task

`rspec-default_http_header` - this gem has been added which allows not to repeat the auth headers in the specs by way of `DefaultHttpHeader` in `spec/rails_helper.rb`

`ParticipationRecord` - this is separate to the existing `EarlyCareerTeacherProfile` because: 1) participants includes mentors (no model for these yet) and 2) separation of concerns.

Bearer tokens vs JWT - Other services within Teacher Services are already using bearer token auth in like this. JWT was also considered but there was nothing in what we are doing here that required it.

API versioning - We had originally written this new API to use headers for API version selection, but after discussion with the cross-team technical architect (Matt Smith) this patch is now using URI based versioning. ADR to follow from Matt. In short it was to keep consistency with other endpoints that providers will already be interacting with, possibly from the same systems.

PR feedback has been dealt with and squashed into this commit because
this project is currently denying merge commits to `develop`.

Co-authored-by: Pavel Lisovin <pavel.lisovin@digital.education.gov.uk>
Co-authored-by: Paul McKibbin <pmckibbin@gmail.com>