Add payment calculator
This is a copy paste from https://github.com/DFE-Digital/cpd-payment-calculations/tree/fd30b3ed074cb9029126963984b49452fe38459d

This was previously a separate repository but the wider team have agreed
that we don't currently have a need to split this out into separate
services.

This is a pure lift and shift to make it easy to follow history with no
integration with the parent project. Full integration can follow with
further commits/PRs.

The specs for the calculator can still be run within the added
subdirectory but haven't yet been added to CI.