Allow multiple params to be passed in Initialize with Config and add in default_config.
Initialize with config was used to pass in the configuration information to the payment calculator.
Now that the bulk of that is in the Contract, and the number of participants and the event type are
going to be passed, it would be good to change it to allow for named parameters to exist as part of
the call.

Additionally, the OpenStruct conversion was really unneccessary since the new methods could be called
directly and having the config stored as a hash meant that it was easier to pass it along the chain
and add additional things to it as required.

This allows for a #default_config to be created to make it more robust against incorrect configuration
and aids in the separation of concerns.

Documentation has been added to make it clearer how to use this as a library class.