Update the payment calculator wrapper to take the contract, number_of_participants and event type to perform the calculations.
The shape of the calculations have had to change from the original test straw-man
which has been brought over from the previous , since this combined events and
contract data into a single input model. This was only used to show progress and
now is superseded by the actual mechanisms using the contract data.