Add a new PaymentService that generates payment html
This method of output is just enough to allow testing a minimum viable output with relevant end users.
It is intended to be called from the rails console and writes html to standard-out.

This intended to be connected to the event aggregation that is a work in progress on another branch to supply real numbers.