New migration to add `early_career_teacher_profiles.uplift` column
This is intended to allow us to capture a School's uplift eligibility at
the point a participant is registered rather than later on when a
payment calculation is made.

The registration capability hasn't been built yet so we can't yet
populate this field.

This field is being put in place to allow us to build the logic for
calculating uplift payments.

Ticket: https://dfedigital.atlassian.net/browse/CPDTP-195
"Aggregate number of participants eligible for uplift payment"