Add a migration that adds uplift columns to early_career_teacher_profiles
These fields are being put in place to support the logic for calculating
uplift payments.

The split of two fields is being maintained here rather than squashing
into a single 'uplift' field because we think we may in future need to
report on uplift payments broken down by reason for uplift.

Although the pupil_premium and sparsity tables are versioned by year,
without this additional copy we would not be able to reconstruct the
values for an early career teacher (ECT) as it was when they registered
if they have moved to another school that has different sparsity/premium.

Ticket: https://dfedigital.atlassian.net/browse/CPDTP-195
"Aggregate number of participants eligible for uplift payment"