Make cohort mandatory for Early Career Teachers
In order to add in uplift payment flags, it was suggested that adding in the cohort default value was a better approach than just co-ercing a value into the database, which could possibly conflict with a null value.
Null values are not a possibility in reality so, as a preliminary step to performing the update, making this a required parameter is a better approach.