Prevent rubocop from autocorrecting version numeric value
Rubocop was enforcing a change to a numerically encoded version to
make it more readable as a number, but less readable as a date.

This is related to the database migration version, so clearer if
laid out as a date format.

Setting this manually and setting a single line disable for Rubocop
prevents the autocorrect from changing it.