Set payment uplift flags on new EarlyCareerTeacherProfile records
Added this in as a callback on the model, but would prefer to
move it to a service wrapper once we have a refactor on the
multistep, which currently leaves either this approach, or
changing the form.save! as options. This was chosen to make
it obvious that it's only for uplift payment.

This is only required the first time a profile is created,
since it shouldn't change even if the underlying school
uplift payment scope changes.