Refactor EarlyCareerTeacherProfile creation to eliminate callbacks on the model
Callbacks are a bit of a code smell in modern Rails applications and can hinder testing etc. with unexpected side-effects.
Refactoring how the AddParticipantForm save! works to make the
creation of EarlyCareerTeacherProfiles a service call and adding
the new uplift values in there makes it cleaner and more testable
for the uplift scenarios without spinning up form testing.