Refactor the Mentor Wizard Create action and add uplift
Add in test to allow us to test the save independently.
Allow it to be called with the mentor_id, to keep it
closer to the EarlyCareerTeachers::Create, but discard
anything additional passed into the initialize.
School Cohort setting information can be pulled into a common shared module to keep the structure a little cleaner.
Add a test on the mentor to enwure shat it creates a profile with the correct uplift.