Strip out the intermediate proxies and reference the EarlyCareerTeacherProfile and MentorProfile directly
It appears that the attempt to avoid the standard polymorphism and go to the
new Rails approach of MultiTable Inheritance is probably not needed when it
really doesn't fix the issue of foreign key integrity since the has_one and
belongs_to are the wrong way round for the association to make this stronger.

After looking at the 4 potential ways of doing this, going back to the one
that we first thought of (i.e. polymorphism directly to the profile tables)
seems to be a better approach.

We can continue to use 'delegated_type', but to the profiles themselves as opposed to the proxies.

Strip out the intermediate proxies and reference the EarlyCareerTeacherProfile
and MentorProfile directly

It appears that the attempt to avoid the standard polymorphism and go to the
new Rails approach of MultiTable Inheritance is probably not needed when it
really doesn't fix the issue of foreign key integrity since the has_one and
belongs_to are the wrong way round for the association to make this stronger.

After looking at the 4 potential ways of doing this, going back to the one
that we first thought of (i.e. polymorphism directly to the profile tables)
seems to be a better approach.

We can continue to use 'delegated_type', but to the profiles themselves as opposed to the proxies.

Remove 'ParticipantDeclaration' from random_record test, since it needs additional traits for the creation.