Switch to not having an intermediate profile declaration for ect or mentors
We made some changes to use delegated_type with intermediate tables and
models to provide some integrity, but that wasn't as robust as hoped, so
dropping those to just use a standard polymorphic set seemed to be the most
sensible approach. This commit flattens the multiple layers down to
profile_declarations -> ect or mentor profiles.