Switch ParticipantDeclaration to UserID
This is a step to keeping the participant_declaration exactly as
submitted to us. They pass us the participant_id (user_id), so
we save that in the record. That way, the event source is as
supplied and not processed in any way that changes the fields.

Since we're accepting more than just the EarlyCareerTeacher, this
would have to have changed to being either a polymorphic association
and since it's only supplied as a participant_id, it didn't make
sense to have it as anything different. The profile_declarations
will hold the actual profile pointed to at the point of declaration.