Switch Profiles to associate with the Profile Declaration
This step completes the declaration piece and allows submitted
declarations to populate the ParticipantDeclaration and
ParticipantProfile. Note that it's not changing the aggregation etc.
yet, but the early_career_teacher_records will continue to
aggregate in the same way as before.

Switch Profiles to associate with the Profile Declaration

This step completes the declaration piece and allows submitted
declarations to populate the ParticipantDeclaration and
ParticipantProfile. Note that it's not changing the aggregation etc.
yet, but the early_career_teacher_records will continue to
aggregate in the same way as before.