Drop the truncate after discussion about migration vs console operations
Simplify the check to use the existing user.participant? call rather than the profile_type.

This change will force a requirement for manual deletion of data in the
ParticipantDeclarations table on the sandbox (and potentially if any
test data exists in other development systems).

There is no relevant data in this table and anything already
there will need to be deleted manually as a console operation.