Refactor RecordParticipantDeclaration
Delegates to user model for queries and profiles.
There was some duplication in this method that was unnecessary, since similar actions were being performed on the user model.
This changes to delegating to there instead.
The factory was also unneccesary, since the profiles could also be queried from the user model.