Update schema validation to match swagger changes for 0.3
Add in course_type migration.

A little bit back-to-front. This was actually used in conjunction with the
docs to generate the schema used for swagger generation and an attempt was
made to use the swagger schemas directly, but after looking at a couple of
approaches, none of the others worked in quite the way that we needed.

I think there's probably a better approach, but continuing to check against
the json schema documents held here continues to work correctly for the
nested types, which it didn't for the other two tried.

Extracted the subschemas into separate components.
Remove actual_lead_provider.nil? guard, since that should never be the case and we seem to have some bad data that needs to be fixed.