Update swagger for new 0.3 schemas
Bring the schema in line with the changes
required for the 0.3 version that we're adding
in the course_type for, and also updates to
the overall shape to bring it in line with the goverment standard: data: {type: <model>, attributes: <list of attributes>}.
This will be a breaking change, since older versions will not adhere to this structure.

In addition, this puts the groundwork in places to enable NPQ declarations to be added in later.