Update the code and tests to support the 0.3 structure.
Since the layout and default schema location has changed, this will apply
the documentation and validation changes to let the code work with the
new json schema shape.

Note that this does not support all of the schema variants, it only uplifts the current ECF early-career-teacher and mentor which can already be submitted to support the new shape of the json, and adds in the required course_type to the participant_declarations table.

Co-authored-by: Pavel Lisovin <pavel.lisovin@digital.education.gov.uk>