Switch the comparison to a direct array comparison
This is a test to actually check that all of the items in the
first set are included in the 2nd set. To do this, you don't
have to take an element and compare it for inclusion in the
other manually, since all of the array arithmetic checks do this.
For elements in A that are not in B, (A-B) will return any elements in A not included in B.