Split the schema into ECF and NPQ specific schemas.
Although there are a lot of similarities, it was decided that it would be better to split early
so that the declarations could be made as ECF or NPQ declarations by the recorder. Since the
recorder performs the schema checking, this will enable it to pick up just the schema required
for that declaration (course_identifier) type.