Convert the declarations controller to allow for a factory based approach to declarations.
Make the declarations more atomic, and eliminate the generic "checking code". The only thing
that keys a declaration type is the "course_identifier".
If this is recognised, the correct "record_declaration" namespaced service is called. This service
should know all about how to record a declaration for that type, including schema validations and
checking that the profiles exist etc. before writing the record.