Eliminate the nil check from validate_provider!
The old code allowed a lead_provider to be nil if a declaration were submitted before the lead_provider was associated with the user profile. This was erroneous and this patch fixes it and updates the tests to require a partnership. It also eliminates an extra 200 test from the API specification, since those tests are included in the request and as the swagger spec is mainly to generate a schema and documentation, they aren't as useful to have them in there.