Implement participant-declarations for the remaining payment declaration types.
This rounds off 'retained-x' and 'completed' declarations for ECF and NPQ.

Schemas validations are now separated into the 'started' (and completed) and 'retained' sections in the validator to make the error messages more specific to the action taken.

Loading one of the specific schemas is now a class level method specific to each of the particular schemas used.