Put a plain English replacement for Z2H heading in readme
It seems from the PR comments that Z2H is an acronym for Zero-to-Hero,
which is a slang way of getting set up quickly.

The readme should be in plain English as much as possible because there
will be many people of varying skills and culture who wish to understand
this repo and how to use it. It certainly threw me, I assumed it was a
mistake.