Delete flapping randomized feature flag tests
Tests should be repeatable. For reasons that are not explained in the
history these tests use a randomized set of factories, as a result this
test file is randomly failing causing a drag on other development. This
should never have been allowed into the main branch.

In order to restore sanity to development this patch deletes the
offending randomization and the tests that depend on it. I shall leave
it to the author of these tests to re-do the coverage in a more stable
way.