[CPDEL-788] Fix cohort lookup in user serializer
The previous method shouldn't be used any more as it uses attributes that aren't
the source of truth. We should look at the school cohort to find the user' cohort.

Also use the find_school_cohort method and amend the includes in the controller to
avoid N+1 lookups to find the cohort related to the school cohort.