Refactor and clean up the broken validation catch at the controller.
The global catch in the controller was incorrectly returning "participant_id" does not match "course_id" when unexpected things were happening (e.g. participant_id did not equate to a user).
This refactor is both a fix for that, and a preparation for a simplification of the common areas between the participant-declarations and the participants API calls.