Track sending of year2020 invites
Because schools will have varying buckets that they fit in, ie the number of schools that have all validated ects is always changing. We need to keep track of which ones we have sent to ensure we don't miss any.

All the remaining sends will go out to schools with induction coordinators that are eligible and not opted out, and this is the first email of this category.

It is impossible to test that this works idempotently due to the way the mail recorder works (breaks half the test suite). Will need to test this manually