Parallelize Cypress and Rspec
Parallelize Cypress
There is a build in parallel functionality with cypress:
See https://github.com/cypress-io/github-action#parallel

But we can't use this without paying for a cypress.io account
(originally tried it and we blew through the limit of free usage in
1 day. Instead, add a new command that randomly splits feature files
based on the number of nodes available. We get diminishing
returns the more nodes we add, so 6 seems like a good sweet spot.

Parallelize rspec

Adds a new command that randomly splits tests randomly, similar to
cypress

Additional js improvements:
 - Improve yarn install command
    use flags based off the node heroku buildpack:
    https://github.com/heroku/heroku-buildpack-ruby/pull/892#issuecomment-640900787
    This seems to prevent the occasional times that yarn was
    just ignoring the cache and taking 5 minutes to install
    rather than 30s

- Remove existing caching steps
   These are already covered by the action we are using

- Remove yarn installation step
   Yarn is preinstalled

- Remove Misc step definition meta tests
  This is a test for a single step that isn't used in any test and takes ages to run