Ignore pundit errors in sentry
We have this line in application.rb
```
config.action_dispatch.rescue_responses["Pundit::NotAuthorizedError"] = :forbidden
```
Which renders a 403 error page when a pundit error is unhandled. However we are still getting errors reported in sentry https://sentry.io/organizations/dfe-bat/issues/2734501640/?project=5748989&query=is%3Aunresolved

This is because the sentry middleware catches the error (and re-raises it) before action dispatch does in the middleware stack. Just ignore this exception in sentry as we know it is being handled.