Fix login sentry error
Some people are hitting the /confirm-sign-in page directly and it is causing a 500 error
https://sentry.io/organizations/dfe-bat/issues/2735066106/?project=5748989&query=is%3Aunresolved
Presumably through a bookmark or browser autocomplete

This is because it finds one of the many users with no login token, and checks their (nil) token expiry. To solve this, just treat the user as if their token is expired if one isn't provided - redirect them to the link-invalid page to sign in again.