Add new sidekiq_worker boxes
We need this because the cloudfoundry healthcheck monitoring of type "process" works where it will fail the healthcheck if all processes spawned from the startup
command die. This means we can't run delayed_job and sidekiq on the same box, as the memory leak causes random killing of processes by the OOM killer. We
can't be in a position where there is either just delayed_job or just sidekiq running during the migration period

Only run one sidekiq process per box - there appear to be 4 physical cpus on each machine, but the vcpu limit that applies is the same. If we just have one process
then there will be two benefits:
   - If a process dies for some reason, the whole box fails the healthcheck and restarts, opposed to just gradually losing processing capacity if there are many
     processes on one box
  - There should be less context switching in the vm so it should be faster

Testing this on a review app and just running random jobs, the sidekiq processes seem to peak at 600MB memory usage, then settle down to 300MB again, so 1024
should be all we need. Add 4 instances on production so we mirror the amount of resources being consumed - the amount of vcpu credits is proportional to
the amount of memory requested so we won't be using 4 times as much cpu https://lists.cloudfoundry.org/g/cf-dev/message/7905

Finally, reduce the number of DJ workers in review/dev from 4 to 3. The addition of sidekiq to these boxes was causing common OOM killing of the sidekiq process
(Everything runs on a single box in those environments)