Fix 404 API responses being html
On some occasions we would return an empty body with text/html, other times json.

This changes the api controller to rely on the exceptions app - see config.exceptions_app = routes in application.rb and ErrorsController, for 404s and 500 errors.
It isn't so straightforward things like BadRequest as we catch a few questionable exceptions to return that, along with a custom error message.