Fix call to update analytics from ECT create service
Previously using `@participant_profile` which was nil, and the spec was just asserting that the job was called.

This call is inside a transaction, so if we can't update analytics then the whole thing gets rolled back.

On production after this is deployed, manually create a Analytics::UpsertECFParticipantProfileJob for all ParticipantProfile::ECTs created after 1pm (we only have one
error, so far)