Don't retry sentry jobs
We ran out of memory in redis because there were around 500 jobs that kept failing because their payload was too big and they were taking up around 1.6GB of memory. There is no point retrying these jobs as they can never succeed and will cause redis to run out of memory.

I also noticed when testing this that we were getting an error in the logs to do with sentry and active job failing to serialize

```
2021-12-15T11:18:40.761Z pid=77624 tid=1tu0 class=Sentry::SendEventJob jid=e09ec3d79d62c44449fa2f13 ERROR: Failed enqueuing Sentry::SendEventJob to Sidekiq(default): ActiveJob::SerializationError (Can't serialize a Hash with reserved key "_aj_symbol_keys")
2021-12-15T11:18:40.761Z pid=77624 tid=1tu0 class=Sentry::SendEventJob jid=e09ec3d79d62c44449fa2f13 ERROR: Async transaction sending failed: Can't serialize a Hash with reserved key "_aj_symbol_keys"
2021-12-15T11:18:40.761Z pid=77624 tid=1tu0 class=Sentry::SendEventJob jid=e09ec3d79d62c44449fa2f13 INFO: Sending envelope [transaction] 7ca6dd6571a74c0bb87b4e14d0f298c0 to Sentry
2021-12-15T11:18:40.795Z pid=77624 tid=1tso class=Sentry::SendEventJob jid=27de4dad6f2fc093723237bb elapsed=0.137 INFO: done
```

https://github.com/getsentry/sentry-ruby/issues/511#issuecomment-248310486 this was still sending the event, but just creates needless logging noise. The way to prevent this is to not use ActiveJob for the sentry send event job. Sentry will subclass ApplicationJob if it is defined:
https://github.com/getsentry/sentry-ruby/blob/9303b9ade03e98096feb67784f8df1c129bdd496/sentry-rails/app/jobs/sentry/send_event_job.rb#L3

To get around this, the new job just copies the (very simple) body from the perform method from Sentry::SendEventJob:
https://github.com/getsentry/sentry-ruby/blob/9303b9ade03e98096feb67784f8df1c129bdd496/sentry-rails/app/jobs/sentry/send_event_job.rb#L25
 opposed to subclassing it.