Fix "boolean" bug with participant declaration api docs
There was a bug where the value for the voided field in the api example was "boolean", when it should be true or false. This is due to a bug in the govuk_tech_docs
renderer where it does something like
```
rendered_value = example.value || example.type
```
And because the value of the example was false, it renders the type as a string, which makes no sense. I tried to fix this in the renderer itself, but the test suite is
entirely broken because it doesn't run as part of CI, so this is a workaround.