Add concept of previous participants to participant band
When calculating output/service fees, we want each band for each milestone type to be calculated using all declarations from that cohort, not
to reset for each month. So we need to pass in the total number of previous declarations in that band so we can calculate where the current
ones lie.