Temporarily remove banding tracker test.
It's seems to be taking over an hour to run under some conditions.
It something on CI is making it not run correctly.
My hunch is that it maybe related to running the tests in parallel but not
having enough connection in the pool or connection timing out.
This will need a bit more investigation but as of now we don't have the
exact (with the seed nuber use to run the tests)
command that run by CI to reproduce this