Move transferring out ppts into active_induction_records for coc_set_participant_categories
This is a temporary change! At the minute when a ppt is transferred an old sit will see them in the transferring out section of the dashboard because we set an end date on their induction record. We don't want to show the transfer out yet. A flag will need to be added to a transferee to say whether the old school has actioned the transfer. This can be done when the transfer out journey is built. For the minute they'll remain under whichever heading is relevant for them.