Bugfix: check for npq profile in admin/participants/show
The previous way of checking for the participant was using legacy associations and it looks like that not all users have this, i.e. an early career teacher profile or mentor against the user. The previous call was causing it to check for the ECTPending view, which has been removed from admin/participant/details component. Doing the check directly on the participant profile should stop this error.