Add additional restriction for sit adding themselves
There was a sentry error showing again that the mailer had failed because of a incorrect template match. I think this is because they've gone through the journey fully rather than just adding themselves via the 'add yourself as a mentor'. This will stop the email being sent if they do go through the full journey.