Check only first name matches when checking for existing validation data
We were checking against a full name match when searching for ecf validation data in the transfer journey. This was causing problems when people were transferring ppts as this is the name stored from the DQT. If there's any difference in this then it won't match. Reducing it to just a first name matches how we initially validate against the DQT.