Amend induction record transfer out scope.
We have to account for scenarios where a ppt is the transferred in but the old school doesnt know. The old school shouldn't be notified of this until the end date has passed. In this case where they dont transfer them out, the ppt will move over to the transferred section of the dashboard. Until that date they will stay in the sit_unknown_transferring_out scope.