Bug-fix prevent a SIT from transferring a user without a previous cohort
There's a bug (https://sentry.io/organizations/dfe-teacher-services/issues/3353853344/?project=5748989&referrer=issue_alert-slack) where if a SIT can try to transfer a ppt from to the 2021 to their school, even if they dont have a school cohort for that year. This blocks them from doing it. We need them, at the minute, to create a full induction programme before they're able to do so. This is a quick fix, as agreed with Mili, Brian and Jo, that tells the user to go to support.