Add email to send to schools whose SITs email has bounced
We sent a batch of comms in May asking school induction tutors to register a programme for the 2022 cohort. We had a number of email addresses that bounced back. We have a specific template to contact all the schools to check that the SITs email is still correct or update it if necessary