Add initializer that sets Google creds as env vars
The Google Drive API can operate in two modes, by using an credentials
file or by reading the values from environment variables.

Adding a new credentials file to the mix will add complication to the
deployment routine so we're taking the env var route. As a result we
need to set the variables in ENV based on the corresponding values from
Rails' credentials.

If no credentials are found nothing will be set.

In order to get the necessary values to try this out you need to set up
a service account on Google Cloud

https://console.cloud.google.com/iam-admin/serviceaccounts/

Then generate a key and set the appropriate variables in your
credentials. For the GOOGLE_DRIVE_NPQ_UPLOAD_FOLDER_ID, pick a folder
on Google Drive and copy the identifier at the end of the URL.

You'll also need to grant permissions on the directory to the email
address given to the service account.