Update existing choose programme spec
This is required as we're now checking for a provider relationship. These are created in the admin console. This means now when a school tries to set up a school cohort, if a provider relationship does not exist it will show an error to the user that the DP no longer works with the LP and they need to select whether they'll continue with the LP.