Use the participant profile state record in the drilldown page
The drilldown page is using the cached training_status value from the
participant profile record to display the participant training status
for NPQ participants, when the API is using the latest participant profile
state record. These two can go out of sync and cause issues and confusion.

Replacing the ParticipantProfile training_status in the drilldown page with the
ParticipantProfile state method ensures the correct state is always displayed.

Eventually the InductionRecord will replace both the training_status and the
ParticipantProfileState. ECF participants already using the InductionRecord,
but the NPQ ones not just yet.