Just log percy snapshot error when build finalized
We have a few flaky tests and occasionally need to re-run part of the
suite. While this isn't ideal, sometimes they pass on the second run but
are reported as failed because Percy can't add new screenshots to a
build that's already been finalized.

In this instance Percy shouldn't cause the build to fail, so we'll
detect finalized build errors and just log them, but `raise` on
everything else.

The error message will still be logged.