Use StringIO instead of tempfile for rack upload
The update to shakapacker brought in a more recent version of rack-test
which changed the behaviour of Rack::Test::UploadedFile[0]. Now it seems
to respect the original filename regardless of what's passed in via the
`original_filename` keyword arg.

We can simplify this by using a StringIO, which has no original filename
so uses the one that's passed in.

[0] https://github.com/rack/rack-test/issues/313#issuecomment-1171473623