Return all cohorts past 2021 when calculating induction records
For an ECF participant, when retrieving induction records, we are
only returning records from the current cohort. The current cohort
is dependant on the time of year if the `multiple_cohorts` flag
has been activited. This means when we attempt to return induction
records for a cohort with milestones in the next year, we are attempting
to query the incorrect cohort record, thus returning a 422 in the API.

Allow multiple cohorts to be queried to return the correct records instead
of only looking at current which could be incorrect.