Sort applications by id to ensure order matches CSV
When inserting records into the csv, we are using find_each
which adheres to the asc ID order. This means that the UIID
of the first record could end up coming in later.

This is causing the test to be flakey due to the order of the
records. Ensure order is respected by sorting by the ID in the
expected CSV.