Ensure we time travel when setting up test scenarios
Scenarios started failing as we've passed the deadline date
on one of the statements (30 September) Although we are time travelling
in the scenarios, we aren't time travelling when creating the statements,
which means we are still attaching the declaration to the latest statement
which keeps shifting as we go forward.

Time travel before a milestone when creating statements to ensure the
dates are consistent. We are no longer using 2 days but 2 months
to ensure declarations fall under the same statement as the tests
have been setup.