Eager load relations used in participants table
This drastically reduces the number of queries on the page by combining
them into a single one with left outer joins.

As ParticipantProfile is polymorphic, we have a number of classes in the
resultset `ParticipantProfile::{ECF,NPQ,Mentor}` but some relationships
were only set up on one of the subclasses. This meant we can't easily
use the same query for the index and would instead of have to treat each
record on a one-by-one basis.

To make this easier I've also moved some relationships from
`ParticipantProfile::ECF` 'up' to `ParticipantProfile`:

* `ecf_participant_eligibility`
* `ecf_participant_validation_data`