Make challenge controller partnership-centric
The partnership controller appeared to be written in a way that assumes
there's only one partnership per cohort per school. I don't think this
is necessarily always the case (it certainly isn't in the seed data).

Now instead of grabbing the partnership based on the school and cohort,
the partnership id is a URL param.

Because multiple partnerships might be available to edit I've added the
relevant details to the top of the challenge form too.