[CPLD-1730] Add serialise for methods to profile
To allow serialising different objects depending on
profile type, add methods to profiles to return the
object we wish to serialise.

With those methods, use new service to defer a participant.
Since we are living in both worlds of new vs old actions,
reuse some of the existing methods for params until we
fully move to the new world where actions will be explicitly
defined