Stop writing Sentry object to country code field
When we can't find a teacher_catchment_iso_country_code we log a message
to Sentry with the country we can't find's name in it. Here the call to
Sentry is the final thing in the method so when we can't find one, the
return value of Sentry#capture_message (a Sentry::ErrorEvent) is set.

This causes an error when writing to the database as the string is
longer than 3 characters.