Ignore pull requests with label "dependencies" rather than actor
Review apps are not created for dependabot prs.

When a pr is merged or closed we attempt to destroy that review app.
The actor performing the closing or merging of a PR is never going to
be dependabot. This means we are attempting to destroy review apps
when they don't exist and fail

Instead look for the label and skip if it is a dependency.