Reduce unfound country sentry messages to warning
The default error level is :error[0], which triggers an alert. In this
case we do want to know about the missing/wrong countries because it'll
show a mismatch between our countries lists in the NPQ and ECF apps, but
it's not important enough to be considered an error.

The originally selected value will still be saved in
`teacher_catchment_country` making the problem rows easy to identify and
fix later if needed.

[0] https://docs.sentry.io/platforms/ruby/usage/set-level/