Explicity define cohort years in school cohorts
The Factory when creating a cohort can create cohorts in years 2055,
which might override the active_registration_cohort.

In the flakey spec we are creating a school cohort using the current year, so
when the validation check kicks in this record isn't returned as the active
registratio cohort is different.

Excplicitly set cohort years so we don't get failures.