Add a rake task for comparing results
We want to compare the results from two queries with production data.
Originally the plan was to add a new endpoint but this was seen as a
risk.

Instead we migrated the actual queries to service objects which
can be called from anywhere.

This sample rake task implements the comparison logic without needing a
new API endpoint.