Use method stubs to avoid relying on Postgres `Date(now())`
When rendering the finance statements page we rely on a scope:
Which uses a Postgres DATE(NOW()) timestamp. This supersedes time
traveling in the specs and returns the latest statement according the time now.

We extract the cohort param from the latest statement, which meant
we were defaulting to 2022. This means we didn't find our November 2021
with that cohort which is why the scenarios fail