Use ActiveRecord range syntax for date queries
This addresses a few problems.

Firstly, general inconsistency across several queries in short
succession.

Secondly, in some of the scopes we are using `DATE(NOW())`, which uses
PostgreSQL's date function. This has the potential to cause problems in
testing when we use TimeCop, as the Ruby-based dates will travel but the
PostgreSQL ones won't.

Thirdly, specs.