Show 'No delivery partner' in email when missing
Currently when sending a fip_provider_has_withdrawn_a_participant email
we try to access the partnership's delivery partner. This is an optional
value so not all partnerships are guaranteed to have one. Currently it
fails when one is missing, here I'm proposing that we instead display a
string of text that says 'No delivery partner' instead.