Refactor NPQ details component, fix last_updated
The first iteration of the NPQ details component accepted the profile
object and traversed along its relationships to get the school,
npq_application and user.

This made testing:

* tricky as more complex setup is needed
* slower as objects need to be persisted by the factory

This refactor changes the component's behaviour so we need to pass in
the school, npq_application and user separately, allowing us to simplify
the component and access values via a single level of delegation.

The `#last_updated` method wsa also tweaked to show the most recent
`#updated_at` from either the participant profile or the user record.