Stop autovalidation on validation details change
Currently we auto-validate a participant when any of their manual
validation data is changed. Once validated it's "fixed" can no longer be
amended by support staff.

This is usually fine but we occasionally have a problem where two fields
need to be amended. After the first one is changed the autovalidation
runs and the second one can no longer be.

To prevent this from happening this change removes the auto-validation.
Now agents will need to click the 'Validate details' button to trigger
it. This way they can be sure all data is correct.