Travel to earliest milestone start date
When attempting to find the next_output_fee_statement for a provider, we are
attempting to get the statement where the `deadline_date` is more than or equal
to the current date. This current date is being set to the first milestone on the
schedule. It seems picking the first one does not respect the start date order,
explicitly set the order to the first date to avoid flakey failures where the statement
cannot be found.