Amend Participant no withdrawn validator
A participant can be reinstated after being withdrawn. However
the validation does not take that into account and finds any
withdrawn participant state for a profile and raises a validation
error.

Change the validation to look for the most recent participant state
scoped to provider, and then check if it has been withdrawn and the
declaration is after the profile has been withdrawn.

Also add tests, we like tests.