Check that user exists before creating declaration attempt
If the participant external ID does not match a user with the
same ID we are failing to create record declarations as it fails
the validation for creating a declaration attempt which requires
a user to exist.

Do not depend on the user having the same ID as the external ID,
look up the user and check they exist before attempting to create
the declaration attempt to avoid unnecessary failures.